+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - 2022-06-30
 
-- add buffs to remind of remaining lives
+- add buffs to remind/warn of remaining lives
+- add xp boost to encourage recently reset players
 
 ## [0.1.0] - 2022-06-30